consolidate all controllers into 1 with methods for each
