--- src/Controller/Admin/AdminUsersActiveController.php
@@ -1,32 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace App\Controller\Admin;
-
-use App\Controller\AbstractController;
-use App\Repository\UserRepository;
-use Symfony\Component\HttpFoundation\RequestStack;
-use Symfony\Component\Security\Http\Attribute\IsGranted;
-
-class AdminUsersActiveController extends AbstractController
-{
-    public function __construct(private readonly UserRepository $repository, private readonly RequestStack $request)
-    {
-    }
-
-    #[IsGranted('ROLE_ADMIN')]
-    public function __invoke(?bool $withFederated = null)
-    {
-        return $this->render(
-            'admin/users_active.html.twig',
-            [
-                'users' => $this->repository->findAllActivePaginated(
-                    (int) $this->request->getCurrentRequest()->get('p', 1),
-                    !($withFederated ?? false)
-                ),
-                'withFederated' => $withFederated,
-            ]
-        );
-    }
-}
